+++ ntoskrnl/io/iomgr/irq.c
@@ -155,7 +155,7 @@ IoDisconnectInterrupt(PKINTERRUPT InterruptObject)
     KeDisconnectInterrupt(&IoInterrupt->FirstInterrupt);
 
     /* Now disconnect the others */
-    for (i = 0; i < KeNumberProcessors; i++)
+    for (i = 0; i < (ULONG)KeNumberProcessors; i++)
     {
         /* Make sure one was registered */
         if (!IoInterrupt->Interrupt[i])
